feat: reusable namespaces
This introduces a new feature to be able to reuse a namespace in
more than 1 Makefile.
In order to support this and to make further development a little
easier the main init file generation has been broken up in a few
different functions and the generation code is now more structure around
the defined Makefiles.
The structure of the generated init file also has changed and there is
now 1 function generated per Makefile and the init method just switches
on which function to call depending on which Makefile was called