fix: clean path to solve crashes on Mac
When launching sagefile with double slashes, e.g. with
`/Users/user/git/repo/subfolder//../.sage/bin/sagefile`, Go crashes when
fetching tools. Cleaning the path by calling it without double slashes
does not produce the crash.

I cannot imagine any reason why it is like this, perhaps a bug to be
filed to Go upstream. The crash reports are very long, so not including
in the commit message.