feat(sglogger): add namespace prefix to logger
This partly reverts 760385e3556f65800ed1bfb9b3642be12b39d923, but adds
the namespace in the prefix separated by a colon.

Examples:
- behavior with this commit:
  `[service:go-lint] linting Go files...'
- behavior before this commit:
  `[go-lint] linting Go files...'
- behavior before 760385e3556f65800ed1bfb9b3642be12b39d923:
  `[service-go-lint] linting Go files...'

This could be useful when running a lot of targets in different
namespaces in parallel, e.g. a CI pipeline running `go-lint` in multiple
services concurrently, to know which namespace the log line relates to.