fix(cache): hierarchically restore Sage cache
If there is a cache miss, we now walk up the tree until we find a hit.
Even if the hit is not a perfect match, it is at least (probably...)
better than a cache miss.

Since this uses a new cache key, it will be a cache miss the first time
after this change is merged.