fix: rollback version update on api-linter
Updates api-linter to version v1.56.1 as v.1.57.0 introduced a new
requirement for (google.api.field_behavior) = IDENTIFIER to be set on
the name field of a proto resource.

There is a conflict between buf lint and api-linter due to an update to
field_behaviour.proto. An issue has been opened buf as fix is on their side:
https://github.com/bufbuild/buf/issues/2474