feat(deps): bump protoc-gen-go-aip-test
This bump adds generation of AIP delete tests if using protoc-gen-go-aip-test.
That means that updating sage to this version can cause build failures in
repositories which do not have code that passes the newly generated tests.

If fixing the business code to adhere to the new tests is not an option,
one can use the `skip` flag in the test suite config to avoid running the
delete tests. See https://github.com/einride/protoc-gen-go-aip-test#skipping-tests
for more information on how to do this.