feat(mdformat): only consider markdown files known to Git
The recursive argument to mdformat (.) is quite slow in repos with lots
of files and folders where no markdown files are expected to be found.
We now instead search for all markdown files that has been checked into
the repo (plus untracked files), instead of letting mdformat do that.

If you want the previous behaviour you have to pass your own arguments
to Command.