fix(sage): remove usage of locally installed balena binary
This override was added in 30ed5e9fe2890ca5ca12eaaf85cc0afb551b0918 to
be able to use balena in NixOS. However, the cost of supporting NixOS is
not worth the issues that arise when everyone potentially runs different
balenacli versions.

This commit removes the check if user already has balena CLI installed,
and will now always use the version specified in this file.