feat(sgterraform): change comment summary to use "destroy" and "add" instead of replace
We have previously used "Replace" to try and make it clear that a
resource was being replaced rather than just being destroyed and/or
added however that can be a little confusing as Terraform itself does
not use any such concept in their plans and it runs the risk of
someone missing the fact that a resource is being destroyed.

So in this commit we instead add the resource to the list of resources
being added and created when a resource is being "replaced".