ci: fix cache thrashing
If multiple workflows call the composite action, the fastest workflow
will always write the cache. This likely means the least amount of
dependencies are cached and all other workflows will suffer and have to
download dependencies not part of the cache.

This change introduces a default cache key equal to the name of the
calling workflow.

For more details on GHA variables, see
https://docs.github.com/en/actions/reference/workflows-and-actions/variables